backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
chromadb==0.4.18
openai>=1.12.0
httpx>=0.25.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
requests>=2.31.0
